Split SettingsScreen into a settings package
SettingsScreen.kt was 1740 lines, and four more files in ui/ existed only
to serve it. Applies the ui/word/ shape, continuing from FavoritesScreen.

The screen splits three ways:

  ui/settings/SettingsViewModel.kt   models, UiState, ViewModel   1053
  ui/settings/SettingsScreen.kt      Screen and Content            404
  ui/settings/preview/SettingsPreviews.kt  the eight previews      324

Four Settings-only files move into the package alongside it:
AboutSection, AppDataSection, VoiceSection and LearningLanguageCard. Each
is referenced by SettingsScreen and nothing else, so leaving them in ui/
would have made the packages mutually dependent — VoiceSection and
LearningLanguageCard both take Settings UiState types as parameters.

SettingsComponents.kt deliberately stays in ui/. It looks Settings-owned
by name, but SectionHeader is also used by DeveloperScreen and
StatsScreen, and LoadingIndicator by WordDetails; only its
DeleteConfirmationDialog is Settings-only. Splitting one dialog out of it
buys less than it costs, so settings/ imports from it instead. The same
goes for StepRow, formatFileSize and the StoreReviewTarget expect/actual
pair, which stay where they are and are now imported explicitly.

This is a pure move. Diffing every moved body against the original files
shows them byte-identical; unlike the Favorites split this one needed no
visibility changes at all, because the previews were the only private
declarations and they moved together.

ui/ drops from 29 files to 25.
