+++ composeApp/src/commonMain/kotlin/com/slovy/slovymovyapp/App.kt
@@ -38,6 +38,7 @@ import com.slovy.slovymovyapp.speech.TextToSpeechManager
 import com.slovy.slovymovyapp.speech.VoiceFilterHelper
 import com.slovy.slovymovyapp.ui.*
 import com.slovy.slovymovyapp.ui.favorites.*
+import com.slovy.slovymovyapp.ui.settings.*
 import com.slovy.slovymovyapp.data.lists.ListsService
 import com.slovy.slovymovyapp.data.lists.WordListsRepository
 import com.slovy.slovymovyapp.data.recovery.RecoverableSense
